Add `__DEV__` expression for dev mode checks
As we begin to implement things like deprecation warnings that we only
want to appear in non-production environments, it's useful to have an
easy check that completely compiles out in production. This commit
adds `babel-plugin-transform-replace-expressions`, which replaces
expressions in code.

The primary expression is `__DEV__`, which compiles to
`process.env.NODE_ENV !== "production"`. When building the browser
bundles, we compile it instead to `false`, so that terser automatcally
removes the entire `if(__DEV__)...` block, resulting in zero runtime
overhead.

This commit also adds the define to Gatsby's webpack config, as Gatsby
doesn't seem to respect the Babel configuration for the main Primer
Components project. In addition, to get Gatsby to run Primer Components
through the new webpack plugin we add, we need to alter the webpack
rules to include files loaded from `@primer/components`.

Finally, when building a production build (like in our browser bundles),
we explicitly set `process.env.NODE_ENV` to `"production"`.